refactor(tmdb): split tmdbService into focused modules
Decompose the 2284-line tmdbService.ts into a tmdbService/ folder,
following the existing storageService convention. tmdbService.ts is now
a thin re-export barrel so all external imports stay unchanged.

Modules: constants, types, httpClient, credentials, cache,
filenameParser, titleMatch, search, multiStrategy, poster, scrape.

No behavior change. tsc --noEmit passes; full backend suite green
(2161 tests).
